pyproject.toml
[project]
name = "django_mqtt_mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "paho-mqtt>=2.1.0",
    "pymysql>=1.1.1",
]
